requirements.txt
streamlit>=1.35.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.5.0
pillow>=10.0.0
numpy>=1.24.0
huggingface_hub>=0.20.0